Refactor pcapng: move SystemdJournalExport to its own file
