+++ src/pcapng.rs
@@ -54,6 +54,7 @@ mod interface_statistics;
 mod name_resolution;
 mod section_header;
 mod simple_packet;
+mod systemd_journal_export;
 
 pub use enhanced_packet::*;
 pub use interface_description::*;
@@ -61,6 +62,7 @@ pub use interface_statistics::*;
 pub use name_resolution::*;
 pub use section_header::*;
 pub use simple_packet::*;
+pub use systemd_journal_export::*;
 
 trait PcapNGBlockParser<'a, En: PcapEndianness, O: 'a> {
     /// Minimum header size, in bytes
@@ -285,36 +287,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     let ts_fractional = (ts % resolution) as u32;
     // XXX should we round to closest unit?
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
-}
-
-#[derive(Debug)]
-pub struct SystemdJournalExportBlock<'a> {
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub data: &'a [u8],
-    pub block_len2: u32,
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, SystemdJournalExportBlock<'a>>
-    for SystemdJournalExportBlock<'a>
-{
-    const HDR_SZ: usize = 12;
-    const MAGIC: u32 = SJE_MAGIC;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], SystemdJournalExportBlock<'a>, E> {
-        let block = SystemdJournalExportBlock {
-            block_type,
-            block_len1,
-            data: i,
-            block_len2,
-        };
-        Ok((i, block))
-    }
 }
 
 #[derive(Clone, Copy, Eq, PartialEq)]
@@ -699,20 +671,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-#[inline]
-pub fn parse_systemdjournalexportblock_le(
-    i: &[u8],
-) -> IResult<&[u8], SystemdJournalExportBlock, PcapError<&[u8]>> {
-    ng_block_parser::<SystemdJournalExportBlock, PcapLE, _, _>()(i)
-}
-
-#[inline]
-pub fn parse_systemdjournalexportblock_be(
-    i: &[u8],
-) -> IResult<&[u8], SystemdJournalExportBlock, PcapError<&[u8]>> {
-    ng_block_parser::<SystemdJournalExportBlock, PcapBE, _, _>()(i)
 }
 
 #[inline]
